requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.8
click==7.1.2
psycopg2-binary==2.8.6
PyYAML==5.4.1
SQLAlchemy==1.3.23
paramiko==2.7.2
asyncpgsa==0.27.1
jinja2==3.0.1
